+++ opentrons-ai-server/Pipfile
@@ -25,7 +25,7 @@ weave = "==0.51.22"
 
 [dev-packages]
 docker = "==7.1.0"
-pytest = "~=7.4.4"
+pytest = "==8.2.1"
 ruff = "==0.4.6"
 mypy = "==1.10.0"
 black = "==24.4.2"
